Restricting Endpoints with Roles.

+++ Controllers/V1/PostController.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using Contracts.V1; 
 using  Tweeting_book.Data;
 using Tweeting_book.Services;
+using Tweeting_book.Extensions;
 
 namespace Tweeting_book.Controllers
 {
@@ -42,7 +43,12 @@ namespace Tweeting_book.Controllers
 
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-          var Post = new Post{Name = postRequest.Name};
+          var Post = new Post{
+
+            Name = postRequest.Name,
+          UserId = HttpContext.GetUserId()
+
+          };
         
           
           await _postService.CreatePostAsync(Post);
@@ -58,14 +64,20 @@ namespace Tweeting_book.Controllers
            [HttpPut(ApiRoutes.Posts.Update)]
             public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatedPostRequest request)
         {
-            var post = new Post
+
+            var UserOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!UserOwnsPost)
             {
-                Id = postId,
-                Name = request.Name
-            };
-            
-            if(post == null)
-            return NotFound();
+                return BadRequest( new{error = "You do not own this post"});
+            }
+
+            var post = await _postService.GetPostByIdAsync(postId);
+             if(post == null)
+            {
+                return NotFound();
+            }
+            post.Name = request.Name;
 
             var updated = await _postService.UpdatePostAsync(post);
             if (updated)
@@ -77,6 +89,13 @@ namespace Tweeting_book.Controllers
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+              var UserOwnsPost = await _postService.UserOwnsPostAsync(postId, HttpContext.GetUserId());
+
+            if (!UserOwnsPost)
+            {
+                return BadRequest( new{error = "You do not own this post"});
+            }
+
             var deleted = await _postService.DeletePostAsync(postId);
             if(deleted )
             return NoContent();
